미들웨어 삭제 및 질문관리 컴포넌트 분리
